+++ src/diagnostic.ts
@@ -1,23 +1,36 @@
-import vscode, { Diagnostic, ExtensionContext, TextDocument } from 'vscode';
+import { ASTRange } from 'greyscript-core';
+import vscode, {
+  Diagnostic,
+  ExtensionContext,
+  Position,
+  Range,
+  TextDocument
+} from 'vscode';
 
 import documentParseQueue from './helper/document-manager';
-
-const hasOwnProperty = Object.prototype.hasOwnProperty;
 
 function lookupErrors(document: TextDocument): Diagnostic[] {
   const errors = documentParseQueue.get(document).errors;
 
   return errors.map((err: any) => {
-    let line = -1;
-
-    if (hasOwnProperty.call(err, 'line')) {
-      line = err.line;
-    } else if (hasOwnProperty.call(err, 'token')) {
-      line = err.token.line;
+    // Lexer error and Parser error
+    if (err?.range) {
+      const range: ASTRange = err.range;
+      return new Diagnostic(
+        new Range(
+          new Position(range.start.line - 1, range.start.character - 1),
+          new Position(range.end.line - 1, range.end.character - 1)
+        ),
+        err.message,
+        vscode.DiagnosticSeverity.Error
+      );
     }
 
+    const firstLine = document.lineAt(0);
+    const lastLine = document.lineAt(document.lineCount - 1);
+
     return new Diagnostic(
-      document.lineAt(line - 1).range,
+      new Range(firstLine.range.start, lastLine.range.end),
       err.message,
       vscode.DiagnosticSeverity.Error
     );